LayoutTests/imported/w3c/web-platform-tests/preload/modulepreload-as-expected.txt


PASS Modulepreload should fire onload with as=""
PASS Modulepreload should fire onerror with as="audio"
PASS Modulepreload should fire onload with as="audioworklet"
FAIL Modulepreload should fire onerror with as="document" assert_equals: Expected onerror to fire expected "error" but got "load"
FAIL Modulepreload should fire onerror with as="embed" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onerror with as="font"
FAIL Modulepreload should fire onerror with as="frame" assert_equals: Expected onerror to fire expected "error" but got "load"
FAIL Modulepreload should fire onerror with as="iframe" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onerror with as="image"
FAIL Modulepreload should fire onerror with as="manifest" assert_equals: Expected onerror to fire expected "error" but got "load"
FAIL Modulepreload should fire onerror with as="object" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onload with as="paintworklet"
FAIL Modulepreload should fire onerror with as="report" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onload with as="script"
PASS Modulepreload should fire onload with as="serviceworker"
PASS Modulepreload should fire onload with as="sharedworker"
PASS Modulepreload should fire onerror with as="track"
PASS Modulepreload should fire onerror with as="video"
FAIL Modulepreload should fire onerror with as="webidentity" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onload with as="worker"
FAIL Modulepreload should fire onerror with as="xslt" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onerror with as="fetch"
PASS Modulepreload should fire onload with as="invalid-dest"
FAIL Modulepreload should fire onerror with as="iMaGe" assert_equals: Expected onerror to fire expected "error" but got "load"
PASS Modulepreload should fire onload with as="sCrIpT"
FAIL Modulepreload should fire onload with as="style" assert_equals: Expected onload to fire expected "load" but got "error"
PASS Modulepreload should fire onload with as="json"
PASS Modulepreload should fire onload with as="text"
